Add GUI to duke
