+++ src/main/java/Duke/SavedTask.txt
@@ -1,0 +1,12 @@
+[E][ ] sad (at: 2020-11-13 18:00)
+[T][ ] ksalfjdskajfksla
+[T][X] jak
+[T][X] dasf
+[T][ ] shfds
+[T][ ] jsakhf
+[T][ ] jsakfdsa
+[T][ ] fjdkslajfkdlsajfdksjfklas
+[T][ ] hjfdklahfdsahkfhdsajkfhdskjafhksdahfkdsahfsad
+[T][ ] dfhsajfhdjskahfjkdsakjfhkjdsahkfhdsakhfkjdaskh
+[E][ ] jdasklfjdskal (at: 2020-12-12 16:10)
+[T][ ] fdhjsak
