+++ src/main/java/Duke/command/DeleteCommand.java
@@ -1,0 +1,43 @@
+package Duke.command;
+
+import Duke.command.Command;
+import Duke.exception.DukeException;
+import Duke.tasklist.TaskList;
+import Duke.utility.Storage;
+import Duke.utility.Ui;
+
+/**
+ * Represents command for delete keyword
+ */
+public class DeleteCommand extends Command {
+    private int index;
+
+    /**
+     * Instantiates a new delete command
+     */
+    public DeleteCommand(int index) {
+        super("delete");
+        this.index = index;
+    }
+
+    /**
+     * Executes the delete command
+     *
+     * @param tasks The list containing all the tasks
+     * @param ui User interface for printing the message
+     * @param storage To write and read from a text file
+     */
+    public String execute(TaskList tasks, Ui ui, Storage storage) {
+        //create new task here
+        try {
+            String message = tasks.deleteTask(index - 1);
+            String output = ui.delete(tasks.numOfTasks(), message);
+            storage.update(tasks.getTasks());
+            return output;
+        } catch (DukeException e) {
+            String output = ui.showLoadingError(e.getMessage());
+            return output;
+        }
+    }
+    
+}
