+++ src/main/java/Duke/exception/DukeException.java
@@ -1,0 +1,14 @@
+package Duke.exception;
+
+/**
+ * Represents exceptions for Duke program
+ */
+public class DukeException extends Exception {
+
+    /**
+     * Instantiates new DukeException
+     */
+    public DukeException(String errorMessage) {
+        super(errorMessage);
+    }
+}
