--- src/main/java/SavedTask.txt
@@ -1,3 +1,0 @@
-[T][ ] slafjkld fhdjaskhf fhakdsfhka hfdk a
-[E][ ] lafj jfdksla jfdkla (at: 2020-10-10 17:29)
-[T][X] jdkslajf
