--- src/main/java/exception/DukeException.java
@@ -1,14 +1,0 @@
-package exception;
-
-/**
- * Represents exceptions for Duke program
- */
-public class DukeException extends Exception {
-
-    /**
-     * Instantiates new DukeException
-     */
-    public DukeException(String errorMessage) {
-        super(errorMessage);
-    }
-}
