+++ src/main/resources/view/DialogBox.fxml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.text.Font?>
+
+<fx:root alignment="TOP_RIGHT" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" prefWidth="400.0"
+         type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1">
+    <children>
+        <Label fx:id="dialog" minHeight="-Infinity" prefWidth="301" text="Label"
+               style="-fx-label-padding: 5 10; -fx-font-weight: bold;
+               -fx-border-color:black; -fx-border-width: 1; -fx-border-style: solid;
+               -fx-border-radius: 16; -fx-background-color: #64d9d3; -fx-background-radius: 16;" wrapText="true">
+            <font>
+                <Font size="14"/>
+            </font>
+            <HBox.margin>
+                <Insets left="5.0" right="10.0"/>
+            </HBox.margin>
+        </Label>
+        <ImageView fx:id="displayPicture" fitHeight="99.0" fitWidth="99.0" pickOnBounds="true" preserveRatio="true" />
+    </children>
+    <padding>
+        <Insets bottom="15.0" left="5.0" right="5.0" top="15.0" />
+    </padding>
+</fx:root>
